Delete unused code.
Delete the unused code for displaying hyper-linked source code for the
standard packages. The data file generator for the feature no longer
works because the generator was written to an old version of go/types.
If the feature is restored, then it should be written from scratch to
take advantage of oracle features.