Add flash messages.
Add support for displaying one-time notification messages on package
pages. This type of message is commonly called a "flash message."

Display a flash message when redirecting to the canonical path for a
package.

Change the package refresh handler to redirect back to the package page
with a flash message on error.