Remove code that blocks mirrors of the Go source
The code was not updated to the new directory tree hierarchy described
in golang.org/s/go14nopkg.

The code blocked repositories that were not mirrors of the Go source
tree. See issue #184. Updating the code to the new directory hierarchy
will likely create more false positives.

The admin tool can be used to manually block a mirror.

Fixes #184.