Link declarations to source
Link all declarations on package pages to the declaration in the source
code. The link is displayed in the top right corner of the <pre> block
on hover over the <pre>.

With this change, package pages now link to const and var declarations
in the source.

Function, type and method headings continue to link to the declaration
in the source.

Bonus change: delete remnants of an experiment from the tdoc SourceLink
method.