Display flash messages on not found page
The reqeust godoc.org/golang.org/x/talks/present redirects with a flash
message to godoc.org/golang.org/x/talks/present. This page is not found.

This change updates the "not found" template to display the flash message.
Previous to this change, the flash message was displayed the next time
the use visited a regular package page.