gosrc: Use unmodified clone URL from import meta tag for cloning.
This fixes a regression in #353 that caused vcs.download to fail for
some valid go gettable packages because an incorrect clone URL value
was being used.

The clone URL value had a possible ".git" suffix trimmed, which is not
correct according to how remote import path resolution is documented at
https://golang.org/cmd/go/#hdr-Remote_import_paths.

Fixes #356.