gosrc: Refactor and document clonePath, repo variables.
Attempt to document and clean up these variables, so that it's more
clear what's going on and why.

Ideally, repo as used previously should be deprecated and removed in
favor of using the repo root import path, which is a more stable and
cleaner value to use as ID for the repo. But that would require a much
more elaborate refactor and likely would be incompatible with the
current data model.