gddo-server: Remove https redirect for "godoc.org" domain
With the move of godoc.org to app engine, the TLS is being stripped
before it hits the gddo-server. This causes an infinite redirect as it
continuously tries to redirect to https even though https is already
being used.

HSTS headers were also removed. Strict TLS is really not needed for
this site. It is a security feature unlikely to provide much security
while at the same time getting in the way if we ever need to host
godoc.org in a non-https enviornment.

Change-Id: Id6f7fd2e1675320ec322077c19834bf4c135ffee
Reviewed-on: https://go-review.googlesource.com/21454
Reviewed-by: Andrew Gerrand <adg@golang.org>