gddo-server: Forbid bot access to import-graph pages.
This CL blocks the access to import-graph pages for robotRequest, which
causes the recent outage since graphs are very expensive for CPU and
disk read I/O. This is a hot-fix that has already been deployed. We'll
work on something permanent later when we come up with some better plan
for robotRequest related issues.

Change-Id: I7e7fbc0538f3d8d26c1ca35cef51e828a79dc1c9
Reviewed-on: https://go-review.googlesource.com/25416
Reviewed-by: Andrew Gerrand <adg@golang.org>