gosrc: Do not exclude cmd/... from standard packages.
This change makes it possible to view documentation for all standard Go
packages, including commands and other packages that match import path
pattern "cmd/...".

That may slightly "pollute" the list of standard Go packages at
https://godoc.org/-/go. But that page is called "Go Standard Packages",
not "Go Standard Libraries", so it's valid to include all standard
packages, including commands, internal, and vendored packages in GOROOT
there. Still, it's likely more useful to rework that page to exclude
these. But it's best to do that separately in a followup issue/CL.

Fixes #437.

Change-Id: Ice05aff3422382246de8a144e240c8efc036c12c
Reviewed-on: https://go-review.googlesource.com/27467
Reviewed-by: Tuo Shan <shantuo@google.com>