gosrc: Fix broken Package Files link.
CL https://golang.org/cl/24513 removed {tag} element from browseURL,
which causes the generated Package Files links to be 404.

Add the {tag} element back, use default branch of the repository.
There is no logic to handle "go1" special case, but neither does
the rest of the logic in getGitHubDir. The scope of this CL is to
fix the broken link only.

Fixes golang/gddo#484.

Change-Id: I34d3c49c0efd98351297272cb41f7a65e1dd47b5
Reviewed-on: https://go-review.googlesource.com/40499
Reviewed-by: Tuo Shan <shantuo@google.com>