talksapp: go-talks.appspot.com => talks.godoc.org
Update references of go-talks.appspot.com to talks.godoc.org, because
the talksapp is hosted on the talks.godoc.org domain now and the old
appspot domain is irrelevant.

Also, some of the references are incorrectly go-talks.appspot.org which
must be a typo, because that domain displays a for sale sign and does
not appear to be owned by Google.

Fixes golang/gddo#532

Change-Id: I8b8ec58e2f2de0e1f2f8d66a6adedc2cb50da754
Reviewed-on: https://go-review.googlesource.com/95256
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>