gddo-server: revert "hide internal packages"
This reverts commit 9b12a26f3fbd7397dee4e20939ddca719d840d2a (CL 90855).

CL 90855 applied a change to fix issue golang/gddo#402, but caused a
regression golang/gddo#533. The regression led to an earlier version
of gddo to be redeployed.

There was an attempt to resolve issue golang/gddo#533 in CL 93196,
which was later picked up and resent as CL 130496.

Tuo and I reviewed CL 130496 and its behavior, and found that
it was not a satisfactory resolution. It introduced additional
complexity and was better in some contexts and situations, but
worse in others. The main downside is that it made it less clear
why some internal packages were being hidden in some places but
not other places.

To fully resolve the original golang/gddo#402, we will need to
come up with and evaluate a more complete plan of action. Until
then, we want to restore the source code to a state which can be
deployed, so take the short term action of reverting the original
CL and go back to a simple and functional behavior: always display
all packages.

Updates golang/gddo#402.
Fixes golang/gddo#533.

Change-Id: I8fb6ac05aa6f118e740e27eaec1bcf2fc9d32766
Reviewed-on: https://go-review.googlesource.com/130937
Reviewed-by: Tuo Shan <shantuo@google.com>