doc: add js/wasm GOOS/GOARCH env pair
Go 1.11 adds an experimental WebAssembly port. The port uses
new GOOS=js GOARCH=wasm values. Some packages, such as syscall/js,
are implemented only for js/wasm environment, so this needs to be
one of the environments to try in order for its documentation to
be available.

Updates golang/gddo#572.

Change-Id: I35b6eae735602e7387beacff07f2b7d6fe323314
Reviewed-on: https://go-review.googlesource.com/131343
Reviewed-by: Tuo Shan <shantuo@google.com>