gddo-server: tee requests to pkgGoDev
At the end of a request to *.godoc.org, an additional request is made
to the teeproxy for pkg.go.dev containing data about the request.

These requests are used to collect data about godoc.org, and ensure that
all incoming traffic to godoc.org can be handled appropriately by
pkg.go.dev.

Change-Id: Ib0b92b81c2a361f5b6a1813ddc62d090edbe4ba1
Reviewed-on: https://go-review.googlesource.com/c/gddo/+/219017
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Andrew Bonventre <andybons@golang.org>
Reviewed-by: Jonathan Amsterdam <jba@google.com>